+++ pycloudflare/services.py
@@ -105,6 +105,9 @@ class CloudFlareService(HTTPServiceClient):
             data['purge_everything'] = True
         return self.delete('zones/%s/purge_cache' % zone_id, json=data)
 
+    def get_ssl_verification_info(self, zone_id):
+        return self.get('zones/%s/ssl/verification' % zone_id)
+
 
 class CloudFlareHostPageIterator(PaginatedAPIIterator):
     page_param = 'offset'